refactor: remove util package, move util package to domain
