+++ src/main/java/coordinatecalculator/domain/X.java
@@ -1,7 +1,5 @@
 package coordinatecalculator.domain;
 
-import coordinatecalculator.util.AxisValidator;
-
 import java.util.Objects;
 
 public class X extends AxisValidator {
